refactor Overflow in TS
