+++ src/Overflow/Overflow.tsx
@@ -1,7 +1,21 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
 import { props, t, ReactNode } from '../utils';
 import ResizeSensor from '../ResizeSensor/ResizeSensor';
+
+export type OverflowProps = {
+  /** react node initially rendered */
+  content: JSX.Element | string,
+  /** react node rendered if `content` overflows its parent */
+  contentIfOverflowing: JSX.Element | string,
+  id?: string,
+  className?: string,
+  style?: React.CSSProperties
+};
+
+export type State = {
+  isOverflowing: boolean
+};
 
 export const Props = {
   content: ReactNode,
@@ -11,15 +25,13 @@ export const Props = {
   style: t.maybe(t.Object)
 };
 
-/**
- * Util component to render a different react node if the original one overflows its parent.
- * @param content - react node initially rendered
- * @param contentIfOverflowing - react node rendered if `content` overflows its parent
- */
+/** Util component to render a different react node if the original one overflows its parent. */
 @props(Props)
-export default class Overflow extends React.Component {
+export default class Overflow extends React.Component<OverflowProps, State> {
 
   state = { isOverflowing: false }
+
+  ref: HTMLDivElement
 
   componentDidMount() {
     this.verifyOverflow();
@@ -29,19 +41,19 @@ export default class Overflow extends React.Component {
     this.verifyOverflow();
   }
 
-  getElementWidth(element) {
+  getElementWidth(element: HTMLElement): number {
     if (element && typeof window !== 'undefined') {
-      return parseFloat(window.getComputedStyle(element).width);
+      return parseFloat(window.getComputedStyle(element).width || '');
     }
-    return null;
+    return 0;
   }
 
   verifyOverflow() {
     if (typeof window !== 'undefined') {
-      const node = ReactDOM.findDOMNode(this.ref);
+      const node = ReactDOM.findDOMNode<HTMLDivElement>(this.ref);
 
       if (node && node.children) {
-        const childrenWidth = this.getElementWidth(node.children[0]);
+        const childrenWidth = this.getElementWidth(node.children[0] as HTMLElement);
         const parentWidth = this.getElementWidth(node);
 
         const isOverflowing = (childrenWidth > parentWidth);
@@ -60,7 +72,7 @@ export default class Overflow extends React.Component {
 
     return (
       <ResizeSensor debounce={10} onResize={this.onResize}>
-        <div {...{ className, id }} style={{ ...style, width: '100%' }} ref={ref => { this.ref = ref; }}>
+        <div {...{ className, id }} style={{ ...style, width: '100%' }} ref={ref => { this.ref = ref!; }}>
           {isOverflowing ? contentIfOverflowing : content}
         </div>
       </ResizeSensor>
