--- src/Overflow/index.js
@@ -1,2 +1,0 @@
-import Overflow from './Overflow';
-export default Overflow;
